Fix: 일정 수정시 스와이퍼 다시 resetting 되게 변경
